requirements.txt
langgraph==0.2.28
langchain==0.3.0
langchain-anthropic==0.3.0
langchain-community==0.3.0
langchain-chroma==0.1.4
chromadb==0.5.3
streamlit==1.39.0
python-dotenv==1.0.1
pypdf==4.3.1
tiktoken==0.7.0
pandas==2.2.2
sentence-transformers==2.2.2